services/fetcher-worker/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
gunicorn==23.0.0

google-cloud-storage==2.18.2

requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
